requirements-EJ.txt
annotated-types==0.7.0
anyio==4.11.0
astroid==4.0.2
beautifulsoup4==4.14.2
black==25.11.0
Brotli==1.1.0
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
ddgs==9.6.1
dill==0.4.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.3.0
isort==7.0.0
lxml==6.0.2
mccabe==0.7.0
mypy_extensions==1.1.0
numpy==2.3.4
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
primp==0.15.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pylint==4.0.3
PyMuPDF==1.26.5
pyparsing==3.2.5
pytest==9.0.1
pytest-mock==3.15.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.8
tomlkit==0.13.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
